requirements.txt
# Core AI and RAG
groq
langchain
langchain-community
langchain-huggingface
sentence-transformers
chromadb
pypdf

# Backend Server
fastapi
uvicorn
python-dotenv
pydantic

# Utilities
requests
typing-extensions